requirements.txt
speedtest-cli
pygame
pyttsx3